// console/src/generated/ProviderDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TlsFingerprintDto } from "./TlsFingerprintDto";

export type ProviderDto = { id: number, name: string, channel: string, settings: unknown, tls_fingerprint: TlsFingerprintDto | null, invalid_tls_fingerprint: unknown | null, tls_fingerprint_error: string | null, enabled: boolean, };